Created new directory for study details and moved relevant components
