+++ src/components/studies/studyDetails/StudyMap.jsx
