+++ src/components/studies/studyDetails/studyMap.css
